requirements65130701931.txt
scikit-surprise 
scikit-learn
pandas

 
